Fix regression of ancestor presentation in Say All
Commit 34c6f7a83fb7 redefined what we considered as "in say all" to
eliminate unwanted focus-mode/browse-mode switches. This redefinition
caused us to stop presenting ancestors, such as the information of a
newly-entered table. Address this by removing the logic to not present
ancestors when in Say All. That should instead be controlled by the
settings related to what context should and should not be presented.

See issue #645.