Adjust when we present label/name for focusable sections
Commit 11683804 fixed a serious chattiness problem with github code
snippets that was caused by a section (generic role) with the focusable
state being given an accessible name by Chromium as a "repair technique"
for bad authoring. That fix, however, apparently broke presentation of
messages in Delta Chat because messages are focusable divs without any
useful ARIA role and without an explicit name. Thus Chromium's repair
technique is kicking in. In order to try to handle both scenarios with
as little pain as is possible, only present the repair-technique-provided
name if the user is in focus mode (i.e. interacting with the content as
an app rather than reading it as a document).

See issue #646.