Have the script manager (not the settings manager) snapshot settings
We need to keep track of runtime settings in cases where a single app has
more than one script associated with it, the most notable example being
terminal apps in which there is a terminal script which handles events
when focus is in the terminal, and the UI toolkit script (e.g. GTK) which
handles events from the non-terminal portions of the app (e.g. menus).
The settings manager was keeping track of runtime settings on behalf of
the script manager. Move that functionality into the script manager, which
is the only consumer with this need.